[!] [Implementation] Increased wished time resolution..
from 1 us to 50 us.

Practically it means that tick is 64 us, not 0.5 us.

Clock is scaled by 1024, not by 8.

And period is 16384 us, not 2048 us.

Why we did this?

When setting prescale speed estimate execution time of time
advancement routine. Here it is significant, near 200 us. It means
that for 2048 us period 10% of time were spent in incrementing
time record.

Main code suffers from it. From it's point of view time sped up.
Because it has less time to think per unit of time.